LayoutTests/editing/deleting/delete-3928305-fix.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/assert-selection.js"></script>
<script>
test(() => {
    assert_selection(
        '<div contenteditable><div><p>foo</p>^bar|<p>baz</p></div></div>',
        'Delete',
        '<div contenteditable><p>foo</p>|<br><p>baz</p></div>');
}, 'Delete characters between blocks in block inserts BR.');
</script>
